Soporta decimales

+++ ejercicio.py
@@ -2,5 +2,5 @@ def sumar(a: int, b: int) -> int:
     return a + b
 
 
-def multiplicar(a: int, b: int) -> int:
+def multiplicar(a: float, b: float) -> float:
     return a * b
